write types

+++ hello.c
@@ -1,4 +1,4 @@
-int write(int fd, const void *buf, int count);
+long int write(int fd, const void *buf, long unsigned int count);
 
 int main(int argc, char *argv[], char* envp[]) {
     write(1, "Henlo World! :3\n", 16);
